[sqlite] add ErrBrokenEngine
